backend/requirements.txt

Django==4.2.7
djangorestframework==3.14.0
django-cors-headers==4.3.1
numpy==1.24.3
scikit-learn==1.3.0
joblib==1.3.2
xgboost==1.7.6
pandas==2.0.3
